backend/requirements.txt
asgiref==3.2.10
requests==2.26.0
django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
PyJWT==2.1.0
python-dotenv==0.21.0
pytest-django==4.5.2
pytest-pythonpath==0.7.4
psycopg2-binary==2.8.6
django-extra-fields==3.0.2
Pillow==9.3.0
djoser==2.1.0
reportlab==3.6.12
django-dotenv==1.4.2